moved state to composables and resolved some type errors
